Fix memory leak
